pyproject.toml
[project]
name = "health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai>=0.8.4",
    "fastapi>=0.118.3",
    "uvicorn>=0.38.0",
]
